case-study/flask-auth/requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-RESTful==0.2.12
Flask-Script==2.0.5
Flask-HTTPAuth==2.3.0
PyJWT==1.0.0
passlib==1.6.2
itsdangerous==0.24
